requirements.txt
opencv-python
numpy
streamlit
Pillow